pyproject.toml
[project]
name = "cdcw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-api-python-client>=2.188.0",
    "gunicorn>=24.1.1",
    "python-escpos>=3.1",
]
